Crate lastpass

An unofficial interface to the LastPass API.

Modules

endpoints
The LastPass API's endpoints.

Structs

Account
A single entry, typically a password or address.

Attachment
Metadata about an attached file.

DecryptionKey
An AES-256 key for encrypting or decrypting things.

Id
A unique resource identifier.

LoginKey
A hex-encoded hash of the username and password.

PrivateKey
A private key that can be used to decrypt items in the password vault.

Session
Information about the current user session.

Vault
Information about all accessible accounts and resources.

Enums

DecryptionError
Errors that are returned when decryption fails.

VaultParseError
Errors that can happen when parsing a Vault from raw bytes.

Constants

DEFAULT_USER_AGENT
The default user agent to use when communicating with the LastPass server.