Module lastpass::endpoints

The LastPass API's endpoints.

Structs

TwoFactorLoginRequired
Two-factor authentication is required.

Enums

EndpointError
Typical endpoint errors.

LoadAttachmentError

LoginError
Possible errors that may be returned by login().

Functions

get_vault
Fetch the latest vault snapshot from LastPass.

get_vault_version
Get the vault version, a number which gets incremented every time account details change (e.g. because you added a password).

iterations
Get the number of iterations to use when encrypting the user's password.

load_attachment

login
Authenticate with the LastPass servers and get a new Session.

logout
Tell the server to invalidate a user's session, logging them out.